Bugfixes in Export/Import All Directories, PermissionRedo

+++ PermissionRedo.ps1
@@ -1,19 +1,28 @@
-$your_principal = "ORGANIZATION.COM\windows-admins"
+function PermissionRedo {
 
-Get-VIPermission -Principal $your_principal | % {
-    
-    $old_perm = $_
+    param([string]$principal)
 
-    $old_entity = $old_perm.Entity
-    $old_role = $old_perm.Role
-    $old_principal = $old_perm.Principal
-    $old_propagate = $old_perm.Propagate
+    if ($principal -eq "") {
 
-    $obj = Get-View -Id $old_perm.Entity.Id
+        echo "Please specify principal!"
+        return;
+    }
 
-    $obj.Name
-    echo "OLD: $($old_role), $($old_principal), $($old_propagate)"
+    Get-VIPermission -Principal $principal | % {
+        
+        $old_perm = $_
 
-    Remove-VIPermission -Permission $old_perm -Confirm:$false
-    New-VIPermission -Entity $old_entity -Principal $old_principal -Role $old_role -Propagate $old_propagate
+        $old_entity = $old_perm.Entity
+        $old_role = $old_perm.Role
+        $old_principal = $old_perm.Principal
+        $old_propagate = $old_perm.Propagate
+
+        $obj = Get-View -Id $old_perm.Entity.Id
+
+        $obj.Name
+        echo "OLD: $($old_role), $($old_principal), $($old_propagate)"
+
+        Remove-VIPermission -Permission $old_perm -Confirm:$false
+        New-VIPermission -Entity $old_entity -Principal $old_principal -Role $old_role -Propagate $old_propagate
+    }
 }
